feat: don't hard-code specific underlying allocator
Right now near-rust-allocator-proxy picks a specific allocator
(jemalloc), which makes it hard for an application (neard) to
independently choose an allocator.

This PR lets the user to choose the underlying allocator.

In particular, the version of jemalloc we are using is obsolete, because
the corresponding rust crate is unmaintained. See

https://github.com/near/nearcore/issues/4766

The reason I am looking into this is that jemalloc doesn't build for me
with clang. I don't know why is that, but upgrading jemalloc is what I
want to try.