chore: bump jemalloc version to 0.5.3
Bumping the version to make it compatible with latest neard.
Also requires a rust version bump because of `tikv-jemalloc-sys`
build commands that rely on newer rust.

Fore more context:
In nearcore, we currently use tikv-jemallocator 0.5.0 which uses
jemalloc version 0.5.3. This makes it incompatible with
tikv-jemallocator 0.4.*

`tikv-jemalloc-sys` 0.5 uses build commands with named arguments in
`println!` which is only supported since Rust 1.58.0.
Bumping to latest Rust to follow the same practice as we do in nearcore.